\name{BayesBDshiny}
\alias{BayesBDshiny}
\title{
Boundary detection and shiny integration
}
\description{
This function produces a local shiny instance for easily producing graphics of boundary detection simulations and real-data analyses.  There is example data and an example function in the "data" folder.  The example data shows how to format image boundary data in a file and was generated using Gaussian intensities and an ellipse boundary.  The example boundary function is a simple function producing a circle boundary.
}
\usage{
BayesBDshiny()
}
\examples{
BayesBDshiny()
}
